DB relationship complete
