+++ src/main/java/com/damyo/alpha/domain/User.java
@@ -1,0 +1,47 @@
+package com.damyo.alpha.domain;
+
+import com.damyo.alpha.domain.Picture;
+import jakarta.persistence.*;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import org.springframework.data.annotation.CreatedDate;
+
+import java.time.LocalDateTime;
+import java.util.ArrayList;
+import java.util.List;
+import java.util.UUID;
+
+@AllArgsConstructor
+@NoArgsConstructor
+@Builder
+@Getter
+@Entity
+public class User {
+    @Id
+    @GeneratedValue(strategy = GenerationType.UUID)
+    @Column(columnDefinition = "BINARY(16)")
+    private UUID id;
+    @Column(name = "name")
+    private String name;
+    @Column(name = "email")
+    private String email;
+    @CreatedDate
+    @Column(name = "created_at")
+    private LocalDateTime createdAt;
+    @Column(name = "profile_url")
+    private String profileUrl;
+    @Column(name = "contribution")
+    private int contribution;
+    @Column(name = "gender")
+    private String gender;
+    @Column(name = "age")
+    private int age;
+
+    @OneToMany(mappedBy = "user")
+    private List<Picture> pictures = new ArrayList<>();
+
+    @OneToMany(mappedBy = "user", fetch = FetchType.LAZY)
+    private List<SmokingData> smokingDatas = new ArrayList<>();
+}
