+++ src/main/java/com/damyo/alpha/repository/InfoRepository.java
@@ -1,14 +1,13 @@
 package com.damyo.alpha.repository;
 
-import com.damyo.alpha.entity.Info;
+import com.damyo.alpha.domain.Info;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
-import java.util.Optional;
 
 public interface InfoRepository extends JpaRepository<Info, Long> {
     Info save(Info info);
-    void deleteInfoBySmokingAreaIdAndId(Long smokingAreaId ,Long id);
-    List<Info> findInfosBySmokingAreaId(Long smokingAreaId);
+    void deleteInfoBySmokingAreaIdAndId(String smokingAreaId ,Long id);
+    List<Info> findInfosBySmokingAreaId(String smokingAreaId);
 
 }
