+++ src/main/java/com/damyo/alpha/repository/UserRepository.java
@@ -1,13 +1,11 @@
 package com.damyo.alpha.repository;
 
-import com.damyo.alpha.entity.Picture;
-import com.damyo.alpha.entity.User;
+import com.damyo.alpha.domain.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Modifying;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.stereotype.Repository;
 
-import java.util.List;
 import java.util.Optional;
 import java.util.UUID;
