+++ src/test/java/com/damyo/alpha/repository/UserRepositoryTest.java
@@ -1,6 +1,6 @@
 package com.damyo.alpha.repository;
 
-import com.damyo.alpha.entity.User;
+import com.damyo.alpha.domain.User;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
